+++ internal/api/web/_app/immutable/entry/start.DvaSc1Mu.js
@@ -1,0 +1,1 @@
+import{o as e,t}from"../chunks/CmXdMWHD.js";export{e as load_css,t as start};